requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
Authlib==1.3.1
backoff==2.2.1
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
dacite==1.8.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.19
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==0.2.0
dataclass-wizard==0.30.1
distro==1.9.0
EditorConfig==0.17.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
fonttools==4.55.3
fsspec==2024.12.0
fuzzywuzzy==0.18.0
grobid-client-python==0.0.9
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
hdbscan==0.8.40
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsbeautifier==1.15.1
kiwisolver==1.4.7
Levenshtein==0.26.1
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multimethod==1.12
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.58.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
validators==0.34.0
visions==0.7.6
weaviate-client==4.10.4
Werkzeug==3.0.6
wordcloud==1.9.4
ydata-profiling==4.12.1
zipp==3.21.0
